requirements_env.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
Jinja2==3.1.3
keras==3.1.1
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyclipper==1.3.0.post5
Pygments==2.17.2
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-image==0.22.0
scipy==1.13.0
shapely==2.0.3
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
validators==0.28.0
Werkzeug==3.0.2
wrapt==1.16.0
